Firebase Notification (background and foreground), open screen along workflow
