+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,6 @@
         </service>
 
         <activity android:name=".basemodel.BaseActivity" />
-        <activity android:name=".mainscreen.MainActivity" />
         <activity
             android:name=".splashscreen.SplashActivity"
             android:label="@string/title_activity_splash"
@@ -31,6 +30,11 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".mainscreen.MainActivity"
+            android:parentActivityName=".splashscreen.SplashActivity">
+        </activity>
+        <activity android:name=".scrollscreen.ScrollActivity"
+            android:parentActivityName=".mainscreen.MainActivity"></activity>
     </application>
 
 </manifest>
