+++ app/src/main/java/com/training/tinhla/training/base/model/constant/ACTIVITY_ID.kt
@@ -1,0 +1,9 @@
+package com.training.tinhla.training.base.model.constant
+
+class ACTIVITY_ID {
+    companion object {
+        val SPLASH = "0"
+        val MAIN = "1"
+        val SCROLL = "2"
+    }
+}
