+++ app/src/main/java/com/training/tinhla/training/base/model/constant/CONSTANT.kt
@@ -1,0 +1,7 @@
+package com.training.tinhla.training.base.model.constant
+
+class CONSTANT {
+    companion object {
+        val EXTRA_ACTIVITY = "ACTIVITY"
+    }
+}
