+++ app/src/main/java/com/training/tinhla/training/scrollscreen/module/ScrollPresenterModule.kt
@@ -1,0 +1,14 @@
+package com.training.tinhla.training.scrollscreen.module
+
+import com.training.tinhla.training.basemodule.PerActivity
+import com.training.tinhla.training.scrollscreen.ScrollInterface
+import com.training.tinhla.training.scrollscreen.ScrollPresenterImpl
+import dagger.Binds
+import dagger.Module
+
+@Module
+abstract class ScrollPresenterModule {
+    @Binds
+    @PerActivity
+    abstract fun presenter(presenter: ScrollPresenterImpl): ScrollInterface.Presenter
+}
