+++ app/src/main/res/layout/activity_scroll.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main_gv"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".scrollscreen.ScrollActivity">
+
+    <include layout="@layout/content_scroll"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
+
+</RelativeLayout>
