+++ app/src/main/res/layout/content_scroll.xml
@@ -6,14 +6,14 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".splashscreen.SplashActivity"
-    tools:showIn="@layout/activity_splash">
+    tools:context=".scrollscreen.ScrollActivity"
+    tools:showIn="@layout/activity_scroll">
 
     <android.support.v4.view.ViewPager
         android:id="@+id/vp_images"
         android:layout_alignParentTop="true"
         android:layout_width="match_parent"
-        android:layout_height="240dp"/>
+        android:layout_height="@dimen/height_bg_header"/>
 
     <ImageView
         android:id="@+id/btn_back"
