+++ app/src/main/res/layout/fragment_background_image.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".splashscreen.header_images.BackgroundImageFragment">
+    tools:context=".scrollscreen.header_images.BackgroundImageFragment">
 
     <ImageView
         android:id="@+id/img_view_background"
