+++ app/src/main/res/values/dimens.xml
@@ -1,3 +1,4 @@
 <resources>
     <dimen name="fab_margin">16dp</dimen>
+    <dimen name="height_bg_header">240dp</dimen>
 </resources>
